fix: initialize member variables to default values
Add in-class initializers (= 0, = nullptr, = false, = {}) to member
variables that were left uninitialized, fixing potential use of
uninitialized memory flagged by cppcoreguidelines-pro-type-member-init.

+++ src/libRd53a/Rd53a2TriggerLoop.cpp
@@ -134,7 +134,7 @@ void Rd53a2TriggerLoop::flexibleTrigger(uint8_t offset, int injDelay, int trigge
     uint8_t trig_start_i = empty_trigs%8; //Index of the first trigger pulse in first frame containing trigger pulses
     int full_frames = floor(float(triggers-(8-trig_start_i)-1)/8); //Number of trigword frames filled with 8 trigger pulses
     uint8_t trig_end_i = (empty_trigs+triggers-1)%8; //Index of the last trigger pulse in the last trigword frame containing trigger pulses
-    for(uint8_t n=0; n<full_frames+2; n++){
+    for(int n=0; n<full_frames+2; n++){
         uint8_t trig_pattern = 0xFF;
         if(n==0) trig_pattern = trig_pattern << trig_start_i;
         if(n==full_frames+1) trig_pattern = trig_pattern & (0xFF >> (7-trig_end_i));
